modularized js file
