+++ index.html
@@ -17,5 +17,5 @@
       </div>
     </div>
   </body>
-  <script src="todo.js"></script>
+  <script src="index.js" type="module"></script>
 </html>
