+++ scripts/elements.js
@@ -1,0 +1,4 @@
+export const addTaskBtn = document.getElementById("addTask");
+export const taskInputText = document.getElementById("taskInput");
+export const taskList = document.getElementById("taskList");
+export const errorMessage = document.getElementById("errorMessage");
