+++ scripts/utilities.js
@@ -1,0 +1,17 @@
+import { taskInputText, errorMessage } from "./elements.js";
+
+export const isUserInputValid = (textContent) => {
+  return textContent.trim() !== "";
+};
+
+export const validateInput = () => {
+  if (isUserInputValid(taskInputText.value)) {
+    errorMessage.innerHTML = "";
+
+    return true;
+  }
+
+  errorMessage.innerHTML = "Please add a task";
+
+  return false;
+};
